pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "epatools"
version = "0.0.1"
description = "ePA Tools"
authors = [{ name = "Ronald Martins", email = "ronald.martins@gematik.de" }]
license = { file = "LICENSE" }
dependencies = [
    "PyYAML==6.0.2",
    "requests==2.32.3"
]

[project.scripts]
epatools = "epatools.main:main"

[igtools.setuptools]
package-dir = {"" = "src"}
packages = ["epatools"]
